dia6/pos_backend/requirements.txt
django==4.2
djangorestframework
cloudinary
python-decouple
djangorestframework-simplejwt
django-cors-headers